+++ src/Stateless/StateMachine.cs
@@ -437,6 +437,14 @@ namespace Stateless
             // Enter the new state
             representation.Enter(transition, args);
 
+            if (_firingMode.Equals(FiringMode.Immediate) && !State.Equals(transition.Destination))
+            {
+                // This can happen if triggers are fired in OnEntry
+                // Must update current representation with updated State
+                representation = GetRepresentation(State);
+                transition = new Transition(transition.Source, State, transition.Trigger, args);
+            }
+
             // Recursively enter substates that have an initial transition
             if (representation.HasInitialTransition)
             {