+++ CHANGELOG.md
@@ -15,7 +15,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 - PR #348 Added trigger parameters to OnTransitioned event
 ### Fixed
-- Fix bug #332 StateMutator is called multiple times with the same state 
+- Fix bug #342 StateMutator is called multiple times with the same state 
+- Fix bug #345 Wrong behavior with FiringMode.Immediate in version 4.3.0
 
 ## [4.3.0] - 2019.11.11
 ### Changed